优化字段类型

+++ models/base_model.go
@@ -1,7 +1,7 @@
 package models
 
 type BaseModel struct {
-	ID        uint32 `gorm:"autoIncrement;primaryKey"`
-	CreatedAt uint32 `gorm:"autoCreateTime;not null;default:0"` // 使用时间戳秒数填充创建时间
-	UpdatedAt uint32 `gorm:"autoCreateTime;not null;default:0"` // 使用时间戳秒数填充创建时间
+	ID        uint  `gorm:"autoIncrement;primaryKey"`
+	CreatedAt int64 `gorm:"autoCreateTime;not null;default:0"` // 使用时间戳秒数填充创建时间
+	UpdatedAt int64 `gorm:"autoCreateTime;not null;default:0"` // 使用时间戳秒数填充创建时间
 }
